% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl_lib.R
\name{convert.vcf}
\alias{convert.vcf}
\title{convert .vcf files to MatrixEQTL genotype input}
\usage{
convert.vcf(vcf.file, which = "GT", map = FALSE, snp.pos = FALSE,
  genotype_file_name = NULL)
}
\arguments{
\item{vcf.file}{vcf.file containing genotypes}

\item{which}{which field of the .vcf file to use: "GT" = genotype,
"DS" = genotype dosage from MaCH/Thunder, "GL" = genotype likelihoods
(default "GT")}

\item{map}{should the map object extracted from}

\item{snp.pos}{names for colums with annotations for custom build,
order: chromosome (chr), length (pos) (default NULL)}

\item{genotype_file_name}{}
}
\value{
either a matrix nsnps x nsample containing genotypes coded as 0,1,2
        or a list with slots "gt.mat", "map", "snp.pos"
}
\description{
convenience function to convert genotypes from .vcf files to the MatrixEQTL format
}
\author{
Ines Assum (2019-03-21)
}
\concept{data conversion functions}
