% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl_lib.R
\name{eqtl}
\alias{eqtl}
\title{basic eQTL interface}
\usage{
eqtl(expression_file_name, genotype_file_name, covariates_file_name,
  gene.position, snp.pos, prefix, redo = FALSE, threshold = 1e-05,
  compute.all = FALSE, what = "table")
}
\arguments{
\item{expression_file_name}{filename of a tab separated file with expression
values in a matrix ngene x nsample}

\item{genotype_file_name}{filename of a tab separated file with genotype
dosage values (usually between 0 and 2) in a matrix nsnp x nsample}

\item{covariates_file_name}{filename of a tab separated file with covariate
values in a matrix ncovar x nsample}

\item{gene.position}{data.frame with gene positions in the first 4 columns
named "chrom", "start", "end", "gene_id"}

\item{snp.pos}{data.frame with SNP positions in the first three columns
named "snp_id", "chrom", "snp_pos"}

\item{prefix}{name prefix for the outputfilenames}

\item{redo}{logical, if TRUE the results will always be recomputed, if FALSE
the results will be loaded if they exist (default FALSE)}

\item{threshold}{significance threshold for eQTLs (default 1e-5)}

\item{compute.all}{MatrixEQTL reports only significant results, if results
for all tests are required set this to TRUE (default FALSE)}

\item{what}{flag to determine what should be returned (default "table"). Can
be set to "object" to get the MatrixEQTL result object.}
}
\value{
if what is "table" the function returns the eQTL results as a table
        if what is "object" the function returns the MatrixEQTL object.
}
\description{
Basic interface to MatrixEQTL passing all data in files
}
\seealso{
Other eqtl functions: \code{\link{eqtl.min.p}},
  \code{\link{eqtl.run}}, \code{\link{manhattan.qtl}},
  \code{\link{myWaitForJobs}}, \code{\link{trans.qtl}}
}
\concept{eqtl functions}
