% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl_lib.R
\name{eqtl.run}
\alias{eqtl.run}
\title{basic eQTL interface}
\usage{
eqtl.run(expr, covar, genotype_file_name, gene.position, snp.pos, id, ...)
}
\arguments{
\item{expr}{expression values in a matrix ngene x nsample}

\item{covar}{covariate values in a matrix ncovar x nsample}

\item{genotype_file_name}{filename of a tab separated file with genotype
dosage values (usually between 0 and 2) in a matrix nsnp x nsample}

\item{gene.position}{data.frame with gene positions in the first 4 columns
named "chrom", "start", "end", "gene_id"}

\item{snp.pos}{data.frame with SNP positions in the first three columns
named "snp_id", "chrom", "snp_pos"}

\item{id}{character indicating an id used for tempfile creation}
}
\value{
the same output as the \code{\link{eqtl}} function
}
\description{
Basic interface to MatrixEQTL passing expression and covariate data in
matrices, and genotypes as a file uses the \code{\link{eqtl}} function
}
\seealso{
Other eqtl functions: \code{\link{eqtl.min.p}},
  \code{\link{eqtl}}, \code{\link{myWaitForJobs}},
  \code{\link{trans.qtl}}
}
\concept{eqtl functions}
