% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl_lib.R
\docType{package}
\name{eQTLpipeline}
\alias{eQTLpipeline}
\alias{eQTLpipeline-package}
\title{full pipeline of eQTL analysis including permutations on HPC systems}
\description{
A package to facilitate the full pipeline of cis eQTL analysis including
permutations on a HPC cluster supported by BatchJobs
}
\details{
\tabular{ll}{
Package: \tab eQTLpipeline\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-11-03\cr
License: \tab LGPL \cr
LazyLoad: \tab yes\cr
}
}
\examples{
#insertexample()
}
\keyword{package}
