% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl_lib.R
\name{find.eSNPs}
\alias{find.eSNPs}
\title{identification of eSNPs}
\usage{
find.eSNPs(expr, covar, gene.position, snp.pos, genotype_file_name, dir,
  threshold, min.perm = 1000, max.perm = 10000, seed = 0,
  exit.criterion = 15, actual.eqtls = NULL, blocksize = 10)
}
\arguments{
\item{expr}{expression values in a matrix ngene x nsample}

\item{covar}{covariate values in a matrix ncovar x nsample}

\item{gene.position}{data.frame with gene positions in the first 4 columns
named "chrom", "start", "end", "gene_id"}

\item{snp.pos}{data.frame with SNP positions in the first three columns
named "snp_id", "chrom", "snp_pos"}

\item{genotype_file_name}{filename of a tab separated file with genotype
dosage values (usually between 0 and 2) in a matrix nsnp x nsample}

\item{dir}{work directory where the BatchJobs registry is stored to. This
path must be accessible by / visible to all compute nodes}

\item{min.perm}{minimal number of permutations}

\item{max.perm}{maximal number of permutations}

\item{seed}{random seed}

\item{exit.criterion}{number of times a more extreme test statistic has to
be observed before a gene is not permuted further}

\item{actual.eqtls}{if the actual eQTLs have already been computed they can
be passed}

\item{blocksize}{the number of permutations collected at the same time. This
parameter tunes the balance between speed and memory usage. The more
permutation runs are collected in the same block, the faster, but
also more memory consuming.}
}
\value{
a table with eQTL P-values and the empirical P-values for eSNPs
}
\description{
Run eQTL permutation analysis using BatchJobs to distribute jobs on a HPC.
This function uses \code{\link{eqtl.run}} function.
}
\references{
GTEx Consortium, Ardlie, K. G., Wright, F. A., & Dermitzakis, E. T. (2015). The Genotype-Tissue Expression (GTEx) pilot analysis: multitissue gene regulation in humans. Science, 348(6235), 648<U+2013>660. \url{http://doi.org/10.1126/science.1262110}
}
\seealso{
Other eqtl HPC functions: \code{\link{find.eGenes}},
  \code{\link{get.egenes.from.esnps}}
}

