% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl_lib.R
\name{get.egenes.from.esnps}
\alias{get.egenes.from.esnps}
\title{Get egenes from a esnp table}
\usage{
get.egenes.from.esnps(esnps, fdr = 0.05)
}
\arguments{
\item{esnps}{an esnps table obtained from \code{\link{find.esnps}}}

\item{fdr}{false discovery rate threshold}
}
\value{
a table of egenes with columns for minimal P-values and emprical
P-values
}
\description{
Get egenes from a esnp table
}
\seealso{
Other eqtl HPC functions: \code{\link{find.eGenes}},
  \code{\link{find.eSNPs}}
}

