% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl_lib.R
\name{get.peer.factors}
\alias{get.peer.factors}
\title{Estimate latent factors using the PEER method}
\usage{
get.peer.factors(k, expr, covar)
}
\arguments{
\item{k}{the number of latent factors}

\item{expr}{expression values in a matrix ngene x nsample}

\item{covar}{measured covariate values in a matrix  nsample x ncovar.
If covar is NULL, no measured covariates will be used.}
}
\value{
the latent factors in a  nsample x k matrix
}
\description{
Estimate latent factors using the PEER method
}

