% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl_lib.R
\name{myWaitForJobs}
\alias{myWaitForJobs}
\title{basic eQTL interface}
\usage{
myWaitForJobs(reg, waittime = 3, nretry = 100)
}
\arguments{
\item{reg}{BatchJobs registry}

\item{waittime}{time to wait before updating job status}

\item{nretry}{number of time to retry getting the job status before throwing #'        an error}
}
\description{
Extension to the waitForJobs function of the BatchJobs package which shows
some strange behaviour when waiting for jobs (database locked)
so we need to make it extra failsafe.
}
\seealso{
Other eqtl functions: \code{\link{eqtl.min.p}},
  \code{\link{eqtl.run}}, \code{\link{eqtl}}
}

