% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl_lib.R
\name{normalize.quantile}
\alias{normalize.quantile}
\title{Quantile normalization}
\usage{
\method{normalize}{quantile}(x)
}
\arguments{
\item{x}{ngenes x nsamples matrix to be normalized}
}
\value{
quantile normalized matrix
}
\description{
Quantile normalization
}

