% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl_lib.R
\name{qquniform}
\alias{qquniform}
\title{Quantile - quantile plot against the uniform distribution}
\usage{
qquniform(p)
}
\arguments{
\item{q}{P-values}
}
\description{
Quantile - quantile plot against the uniform distribution
}

