% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl_lib.R
\name{trans.qtl}
\alias{trans.qtl}
\title{transQTL interface}
\usage{
trans.qtl(prefix, genotype_file_name, vcf = FALSE, snp.pos = NULL,
  expression_file_name, covariates_file_name = character(),
  threshold = 1e-05, compute.all = FALSE, what = NULL,
  min.pv.by.genesnp = FALSE, save.memory = FALSE, verbose = TRUE,
  redo = TRUE, load.qtls = FALSE, keep.outfile = FALSE,
  id.check = FALSE, gene.position = NULL)
}
\arguments{
\item{prefix}{path and prefix for the outputfiles (default NULL)}

\item{genotype_file_name}{filename of a tab separated file with genotype
dosage values (usually between 0 and 2) in a matrix nsnp x nsample}

\item{vcf}{is the genotype file a vcf file that needs to be converted?
(default FALSE)}

\item{snp.pos}{data.frame with SNP positions to be added to matrixEQTL object
in the end. Columns named "snps", "chr", "pos", containing
information in that order (default NULL)
if vcf=TRUE and snp.pos=TRUE | snp.pos=NULL, information is extracted
from vcf file unless save.memory=TRUE.
if snp.pos=FALSE, no annotation is added}

\item{expression_file_name}{filename of a tab separated file with expression
values in a matrix ngene x nsample.
if expression_file_name is a numeric vector (lenght nsample),
matrix (ngene x nsample) or data.frame (ngene x nsample), the
expression file is created with the name prefix_gene.txt.}

\item{covariates_file_name}{filename of a tab separated file with covariate
values in a matrix ncovar x nsample
to not use any covariates, set covariates_file_name=character()
if covariates_file_name is a numeric vector (lenght nsample),
matrix (ngene x nsample) or data.frame (ngene x nsample), the
covariate file is created with the name prefix_covariates.txt.}

\item{threshold}{significance threshold for eQTLs (default 1e-5), set to 1 to
get all QTLs. threshold set to 1 for compute.all=TRUE}

\item{compute.all}{MatrixEQTL reports only significant results, if results
for all tests are required set this to TRUE (default FALSE)}

\item{what}{flag to determine if a table or the MatrixEQTL object should be
returned (default NULL). If unspecified, a table is returned except
either save.memory=TRUE or min.pv.by.genesnp=TRUE, then the MatrixEQTL
result object is returned.}

\item{min.pv.by.genesnp}{option to report the minimal pvalue per SNP and per
gene in the MatrixEQTL object (default FALSE). If you want to use this
option, make sure to set what="object" to get the results.}

\item{save.memory}{prevent FDR correction (default FALSE). only functional in
the github-version of MatrixEQTL, see also
https://github.com/andreyshabalin/MatrixEQTL/issues/8}

\item{verbose}{print comments of progress (default TRUE)}

\item{redo}{logical, if TRUE the results will always be recomputed, if FALSE
the results will be loaded if they exist (default TRUE)}

\item{load.qtls}{(default FALSE) only relevant for save.memory=TRUE which is
mostly used for big datasets. In that case, QTL results are not
reported in the MatrixEQTL object but can be read in with
load.qtls=TRUE.}

\item{keep.outfile}{unless compute.all=T or save.memory=T, results are written
       to a temp file that is deleted. (default FALSE)

Parameters /options I would like to add:}

\item{id.check}{check if ids and order of ids match in genotype, expression
and covariate data (default FALSE)}

\item{gene.position}{data.frame with gene positions in the first 4 columns
named "gene_id", "chrom", "start", "end"}

\item{threshold.trans}{Parameters/options in the old wrapper that do not work currently:}
}
\value{
if what="table", the data.frame contained in the MatrixEQTL object
        me$all$eqtls is returned. If what="object", the whole MatrixEQTL
        object is returned.
}
\description{
Interface to the MatrixEQTL Matrix_eQTL_engine()
This is a simplified version for more general (trans)QTL analyses,
e.g. GWAS or metaboliteQTL and does not need gene annotations
}
\seealso{
Other eqtl functions: \code{\link{eqtl.min.p}},
  \code{\link{eqtl.run}}, \code{\link{eqtl}},
  \code{\link{manhattan.qtl}}, \code{\link{myWaitForJobs}}
}
\author{
Ines Assum (2019-03-21)
}
\concept{eqtl functions}
